+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changes
+
+- Change subcommand `list` to `log`.
+- Change subcommand `tasks` to `list`.
+
 ## [0.4.0] - 2021-05-04
 
 ### Changed